requirements.txt
# Flask and web server related
flask>=2.1.0,<3.0.0
Flask-Cors>=3.0.0,<4.0.0 # Optional, but good to have for dev
requests>=2.28.0,<3.0.0

# Core ML/Data Libraries
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0 # If needed for data manipulation later
matplotlib>=3.5.0,<4.0.0

# Image Processing
Pillow>=9.0.0,<11.0.0
opencv-python>=4.6.0,<5.0.0
scikit-learn>=1.0.0 # Added for K-Means
scikit-image>=0.19.0 # Enhanced image processing

# Advanced color analysis
colormath>=2.1.1 # For accurate color operations
webcolors>=1.12 # For named colors
colour-science # For color space conversions
kneed>=0.7.0 # For finding optimal k in elbow method

# Deep Learning Frameworks & Models
tensorflow>=2.10.0,<2.17.0 # MODIFIED: Widened TF version range for compatibility
torch>=1.12.0,<3.0.0 # Allow PyTorch 2.x for Python 3.11 compatibility
torchvision>=0.13.0,<1.0.0 # Keeping torchvision below 1.0 for now
timm>=0.6.0,<1.0.0
transformers>=4.20.0,<5.0.0

# Utilities
tqdm>=4.64.0,<5.0.0

motor>=3.0.0 # For async MongoDB
pymongo>=4.0.0 # For sync MongoDB
aiohttp>=3.8.0 # For async requests
aiohttp_socks>=0.7.0
aiofiles>=0.8.0
uvloop>=0.17.0; sys_platform != 'win32' # Ignored on Windows anyway
aiohttp_retry>=2.5.2
stem>=1.8.0
certifi>=2022.12.7
